Address review notes
